Current Prototype for the Project
Has the ability to upload images when the server is active, can add notes using the "add note" functionality, and allows for the appending of a note's information to the screen when its corresponding button is clicked on. Saving/loading is still a work in progress, but some groundwork has been laid to work towards that.